Further simplification.
It turns out that MemberType.Event does not need to be considered at
all.  There is an autogenerated backing field (i.e. MemberType.Field)
for each event.

Also made a new type to represent an individual search, and placed the
scanned object set and the stack of fields in there.  This allows us to
remove the "avoid concurrent runs" checks.